Move DependencyResolver to Distribution.Solver.* namespace
