+++ cabal-install/Distribution/Client/Dependency/TopDown.hs
@@ -22,9 +22,6 @@ import Distribution.Client.Dependency.TopDown.Constraints
 import Distribution.Client.Types
          ( UnresolvedPkgLoc
          , UnresolvedSourcePackage, enableStanzas )
-import Distribution.Client.Dependency.Types
-         ( DependencyResolver )
-
 
 import qualified Distribution.Simple.PackageIndex  as InstalledPackageIndex
 import Distribution.Simple.PackageIndex (InstalledPackageIndex)
@@ -54,6 +51,7 @@ import Distribution.Text
 
 import           Distribution.Solver.Types.ComponentDeps (ComponentDeps)
 import qualified Distribution.Solver.Types.ComponentDeps as CD
+import           Distribution.Solver.Types.DependencyResolver
 import           Distribution.Solver.Types.InstalledPreference
 import           Distribution.Solver.Types.LabeledPackageConstraint
 import           Distribution.Solver.Types.PackageConstraint
